add react router and probe details page WIP
