+++ package.json
@@ -7,6 +7,8 @@
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-refetch": "^2.0.3",
+    "react-router": "^5.0.0",
+    "react-router-dom": "^5.0.0",
     "react-scripts": "2.1.8"
   },
   "scripts": {
